Restructures components, views
